+++ CHANGELOG.md
@@ -4,6 +4,15 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
 ## [Unreleased]
 ### Changed
+- `JSONAuto` renamed to `JSONDeserializer`
+- `JSONAutoTest` renamed to `JSONDeserializerTest`
+
+### Added
+- `JSONAuto`: new version (pass-through to `JSONDeserializer`)
+- `DummyClasses` (test): split out from `JSONDeserializerTest`
+
+## [0.2] - 2019-04-23
+### Changed
 - `JSONAutoTest`: added more tests (derived classes)
 - `JSONAuto`: added code to check for `@JSONName` annotation
 - `JSONAutoTest`: added tests for `@JSONName` annotation